third_party/blink/web_tests/http/tests/security/xss-DENIED-javascript-variations-expected.txt

CONSOLE ERROR: Unsafe attempt to initiate navigation for frame with URL 'data:text/html,<p>Inner text should not be replaced.<p>' from frame with URL 'http://127.0.0.1:8000/security/xss-DENIED-javascript-variations.html'. The frame attempting navigation must be same-origin with the target if navigating to a javascript: url

CONSOLE ERROR: Unsafe attempt to initiate navigation for frame with URL 'data:text/html,<p>Inner text should not be replaced.<p>' from frame with URL 'http://127.0.0.1:8000/security/xss-DENIED-javascript-variations.html'. The frame attempting navigation must be same-origin with the target if navigating to a javascript: url

CONSOLE ERROR: Unsafe attempt to initiate navigation for frame with URL 'data:text/html,<p>Inner text should not be replaced.<p>' from frame with URL 'http://127.0.0.1:8000/security/xss-DENIED-javascript-variations.html'. The frame attempting navigation must be same-origin with the target if navigating to a javascript: url

CONSOLE ERROR: Unsafe attempt to initiate navigation for frame with URL 'data:text/html,<p>Inner text should not be replaced.<p>' from frame with URL 'http://127.0.0.1:8000/security/xss-DENIED-javascript-variations.html'. The frame attempting navigation must be same-origin with the target if navigating to a javascript: url

CONSOLE ERROR: Unsafe attempt to initiate navigation for frame with URL 'data:text/html,<p>Inner text should not be replaced.<p>' from frame with URL 'http://127.0.0.1:8000/security/xss-DENIED-javascript-variations.html'. The frame attempting navigation must be same-origin with the target if navigating to a javascript: url

CONSOLE ERROR: Unsafe attempt to initiate navigation for frame with URL 'data:text/html,<p>Inner text should not be replaced.<p>' from frame with URL 'http://127.0.0.1:8000/security/xss-DENIED-javascript-variations.html'. The frame attempting navigation must be same-origin with the target if navigating to a javascript: url

CONSOLE ERROR: Unsafe attempt to initiate navigation for frame with URL 'data:text/html,<p>Inner text should not be replaced.<p>' from frame with URL 'http://127.0.0.1:8000/security/xss-DENIED-javascript-variations.html'. The frame attempting navigation must be same-origin with the target if navigating to a javascript: url

CONSOLE ERROR: Unsafe attempt to initiate navigation for frame with URL 'data:text/html,<p>Inner text should not be replaced.<p>' from frame with URL 'http://127.0.0.1:8000/security/xss-DENIED-javascript-variations.html'. The frame attempting navigation must be same-origin with the target if navigating to a javascript: url

CONSOLE ERROR: Unsafe attempt to initiate navigation for frame with URL 'data:text/html,<p>Inner text should not be replaced.<p>' from frame with URL 'http://127.0.0.1:8000/security/xss-DENIED-javascript-variations.html'. The frame attempting navigation must be same-origin with the target if navigating to a javascript: url




--------
Frame: 'VictimFrame'
--------
Inner text should not be replaced.
